Az bot: add create v3, v4 bot webapp bot tests, additional refactoring, rename custom -> bot_operations
