+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/__init__.py
@@ -13,7 +13,7 @@ class BotServiceCommandsLoader(AzCommandsLoader):
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
         custom_type = CliCommandType(
-            operations_tmpl='azure.cli.command_modules.botservice.custom#{}',
+            operations_tmpl='azure.cli.command_modules.botservice.bot_operations#{}',
             client_factory=get_botservice_management_client)
         super(BotServiceCommandsLoader, self).__init__(cli_ctx=cli_ctx,
                                                        custom_command_type=custom_type,
