+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/tests/latest/test_bot_commands.py
@@ -40,7 +40,40 @@ class BotTests(ScenarioTest):
         self.cmd('az bot delete -g {rg} -n {botname}')
 
     @ResourceGroupPreparer(random_name_length=20)
-    def test_webapp_bot(self, resource_group):
+    def test_create_v3_webapp_bot(self, resource_group):
+        self.kwargs.update({
+            'botname': self.create_random_name(prefix='cli', length=15),
+            'app_id': '0cc66483-d276-4830-9afd-9fe6b654a255',
+            'password': 'testpassword'
+        })
+        dir_path = os.path.join('.', self.kwargs.get('botname'))
+        if os.path.exists(dir_path):
+            # clean up the folder
+            shutil.rmtree(dir_path)
+
+        self.cmd('az bot create -k webapp -g {rg} -n {botname} --appid {app_id} -p {password} -v v3', checks=[
+            self.check('appId', '{app_id}'),
+            self.check('appPassword', '{password}'),
+            self.check('resourceGroup', '{rg}'),
+            self.check('id', '{botname}'),
+            self.check('type', 'abs')
+        ])
+
+        # download the bot
+        self.cmd('az bot download -g {rg} -n {botname}', checks=[
+            self.exists('downloadPath')
+        ])
+        self.check(os.path.isdir(os.path.join('dir_path', 'postDeployScripts')), True)
+
+        # publish it back
+        self.cmd('az bot publish -g {rg} -n {botname} --code-dir {botname}', checks=[
+            self.check('active', True)
+        ])
+        # clean up the folder
+        shutil.rmtree(dir_path)
+
+    @ResourceGroupPreparer(random_name_length=20)
+    def test_create_v4_webapp_bot(self, resource_group):
         self.kwargs.update({
             'botname': self.create_random_name(prefix='cli', length=15),
             'app_id': '0cc66483-d276-4830-9afd-9fe6b654a255',
